[appmgr] Implement OnTerminated in appmgr.
OnTerminated was introduced to solve CP-82, and will eventually
deprecate Wait (CP-87).

This change implements OnTerminated in appmgr by piping a required
callback thorough the component start and lifecycle flow.

I also introduce FailedComponentController, a ComponentController
that never successfully started. The purpose of this controller is to
bind to the incoming request and respond with an appropriate termination
event.

CP-82: #done
CP-87: #progress

TEST=fx run-test appmgr_unittests

Change-Id: I32fdc7c123a6829877ff22b7273da2423b80fe33