[fidl] Add GN targets for the simple C client bindings
This adds GN build targets for generating and linking against the FIDL
simple C bindings. These can be used by depending on the target:
//path/to/fidl:fidl.library.name_c_client

The generation rules for this library are a bit special since the code
is actually generated by fidlc (the front end) instead of a backend as
usual, so these files are always generated and only compiled in when
the target is depended on. The Simple C bindings are also generated in
the default (target) toolchain, like the C++ and Rust bindings, since
they can be fairly freely mix and matched with other C++ code.

FIDL-184 #comment
Test: Build

Change-Id: Ifba3f1ebc33aec0182dd121313f65faf7f50fd66