[debugger] Improve help text and ID loading.
Adds a tip on how to connect when the debugger starts in interactive
mode. Removes the post-connection help text when not run in an
interactive mode.

Loads the symbol ID file after a successful connection rather than on
startup. This is because the system symbols may have changed since the
last connection (in the case where the debugger stays running but the
system it's connecting to is being updated).

TEST=manual

Change-Id: I3dbae7a71185515efa30539efe1ecb4d719f4cb9