[host][xdcserver] Beginnings of xdc host server.
The xdcserver will sit between host side clients and libusb.

Clients will open a socket to the xdc server and register for a stream
id.

In a future CL, client reads / writes after registering will
be transformed into libusb requests by the xdcserver, which are
then passed on to the xdc server on the device side and routed
to the appropriate stream id.

TEST= out/build-zircon/tools/xdc-server and
      run test app added in
https://fuchsia-review.googlesource.com/c/zircon/+/174059

Change-Id: I2420317a209f42b56bf9968fce08eec3ae1ef775