[build][scripts] Build complete ZBI, boot it in QEMU
The primary outputs of the build are now kernel.zbi and zircon.zbi.
The latter is a complete ZBI for standalone booting.
The run-zircon script now uses zircon.zbi by default.
The deprecated -x switch makes it use the old method with direct
kernel loading rather than a trampoline.

The old names zircon.bin and bootdata.bin are made as links, and the
shimmed qemu-zircon.bin is still made.  These names are deprecated and
will be removed when all the consumers of the build have been updated.

Test: CQ
Change-Id: I37d8661f8f5766e3947d0a6f71b13178a82af3c4