[vim][audio] VIM2/S912 SPDIF/HDMI driver
Add support for SPDIF over HDMI on VIM2.

Note: right now, this HDMI driver is taking full control of all of the
audio hardware, and there is a bunch of HDMI specific code which is
leaking into the SPDIF support.  When the day comes that we have
proper support for composite devices and we no longer have to merge
the DAI component of the driver with the codec, we can come back here
and properly split the two apart.

Right now, the code is mostly (but not entirely) separated, so it
should be pretty easy to split.  Things having to do with format
support, N/CTS calculations, and make/mfr/unique_id, will all come out
of the SPDIF code and become part of the "HDMI codec" half of things,
while all of the functionality for configuring and running the VIM2
SPDIF block should end up in the S912 DAI driver.

Test: Manual on VIM2 hardware
Change-Id: I33a6f5cf9172ec15b2e87839461493b0413ee47e