[USB Video] Moving usb-video driver to garnet
I have moved the usb_video driver from Zircon into Garnet, this should allow to use the FIDL C++ support.

We could eventually migrate this back into Zircon when FIDL C++ support is available in there too.
Having the driver in Garnet will also speed up the co-development of the driver and the client library :)

To get this to work, you need to comment out the content of the rules.mk file in Zircon's usb-video directory.

TEST = Just moved the Zircon driver into Garnet. Removed rules.mk and replaced it with BUILD.gn
       video_test still works fine.

Change-Id: I26af8622179661965c879ce19fe656c266870a2c