[vulkan] Add a loader service for vulkan client drivers
This adds a service for the vulkan client loader (libvulkan.so) to
request the vulkan client driver library for the system. The client
libraries are shared libraries that must match up with the device
driver, so they can't be packaged along with the program that wants to
use vulkan. Instead, a program using vulkan packages the vulkan loader
which queries this service for the library to load.

For now, the loader just opens the library in /system/lib and returns a
VMO handle. In the future, this service will load the library from a
package.

DX-303 #comment
MA-416 #comment

Test: Run programs that load Vulkan

Change-Id: I9dfe66e3aaf61a3a43508e62bbf8792648917f2b