[bt][l2cap] Use flex arrays in SignalingChannel payloads
Use a final flexible array member in structure definitions for
SignalingChannel payloads. This makes the payload structs be the
minimum expected size for the command type, which helps the structs'
use for input parsing.

However, this does complicate the structs' use for building outbound
payloads.

Update Command Reject payload building.

Test: bt-host-unittests
Change-Id: I89170b17e7fc305d45ba29df1c5e91657d1db367