[escher] Introduce RenderQueue and PaperRenderQueue.
RenderQueue is a fairly simple class that supports
generic sorting of render-items before iterating over
them to emit commands into a Vulkan CommandBuffer.
These commands are not generated by the RenderQueue
itself, but rather by callback functions that the
client pushes into the queue along with other data.

PaperRenderQueue builds atop RenderQueue to provide
a simple interface for rendering Escher objects.
The current implementation is a proof-of-concept
that is used in the Waterfall 2 demo; upcoming CLs
will bring PaperRenderQueue to parity with
PaperRenderer.

ES-83 #comment

TEST: Ran 'escher_unittests'.  Added new tests to:
  - mesh_spec_unittest.cc
  - paper/paper_render_queue_unittest.cc
  - renderer/render_queue_unittest.cc
  - util/bit_ops_unittest.cc

Change-Id: I55adc437b0130e58a9bd026145be3345fdc40239