[modular] MessageQueues treat messages as fuchsia.mem.Buffer instead of string
Using a fidl 'string' type for message data is problematic, since data
is not meant to be a blob, not utf8 encoded. Since we don't want to
restrict message sizes, this iteration treats the data as
fuchsia.mem.Buffer. In the future, we may introduce a variant of
MessageSender.Send() or MessageReader.OnReceive() that operates on
vector<uint8>.

* Interfaces in message_queue.fidl use fuchsia.mem.Buffer instead of
  string.
* Introduce MessageQueueClient that is a wrapper around the MessageQueue
  interface and also has helper methods for receiving messages as
  std::string.
* Introduce MessageSenderClient that is a wrapper for MessageSender that
  operates on std::string instead of fuchsia::mem::Buffer.
* Deprecate MessageReceiverClient in favour of MessageQueueClient.
* Update all the tests to use these new interfaces.

Tested=Ran all modular tests (including integration tests).

Change-Id: I5da57bf20d45424724f09cef36f5743efb5c1628