[appmgr] Add config file for scheme mapping.
Currently, the mapping from HTTP(S) -> network component is hard-coded
in CreateComponent. Since we may add other scheme mappings in the
future, it is preferable to specify this policy in a configuration
file.

This CL doesn't yet use the config file anywhere. It only adds the
file and a parser.

CP-69 #progress

TESTED=scheme_map_unittest

Change-Id: Ia198bb7491ab8bc26380397d733fd27e76c33d6b