[perftest] Change test_suite field to match the currently-used name
The current name for this test suite that we're uploading test results
under (on https://chromeperf.appspot.com) has a "fuchsia." prefix.

The "test_suite" field in the perf results JSON file is not used yet,
but let's make it so that when it is used, the results-to-Catapult
converter doesn't have to add a "fuchsia." prefix implicitly.

Bug: IN-442
Test: none
Change-Id: I7271c0436d90b7fe01d173e930f8a5d0da8f1c6b