[mediaplayer] async conversion pipeline builder
This CL makes the previously synchronous 'conversion pipeline
builder' work asynchronously. The conversion pipeline builder adds
nodes to a graph to convert an elementary stream from one type
to another (i.e. it adds decoders). We will soon need to have
decoder creation be asynchronous, so the builder must be async
as well.

TEST: No behavior change

Change-Id: Iec8dc6d4a038e427742a5b49c14f27164baecd5e