[build] Compile debug builds with frame pointers.
Libunwind is not perfect with our build's unwind tables and as a result
stacks from both crash dumps and in the debugger aren't very good.

This change adds full stack frames for target debug builds so the
debugger and kernel stack dumps can work better. Release build defaults
to no frame pointers but it can be toggled on with a build flag.

Hopefully we can fix our unwind tables and default to unwind tables only
for all builds.

ZX-2361

Change-Id: I31c4d44dfc678d99eb51ed0848abf7b547e9cb56