[debugger] Fix passing arguments to programs.
The debug agent was only passing the arguments as the argv to the
launcher rather than the entire command line. This messes up programs
since they expect their own name to be in argv[0].

Updates the documentation for passing arguments to debugged programs.
The parser handles this case already, anything after the first
non-switch parameter doesn't count as a switch, so everything works the
way people expect.

TEST=manual

Change-Id: I7882fded09c03479890e776dd25c806a4ab8c253