[tee][optee] Add secure world shared memory
The shared memory that is used for communication between the secure
world and the non-secure world is provided by the Secure OS. It
allocates a portion of it's RAM, leaving it as non-secure, and then
provides the physical address and size to the REE OS via an SMC call.
This memory range is then split into private (driver communication) and
public pools (client app communication).

For now, we're just using MMIO to map the DRAM as it is currently the
only mechanism we have available to us. This will need a better long term
solution for this type of memory.

Test: Manual testing on VIM2 to ensure we can map the ranges provided to
us and allocate shared memory objects.

Change-Id: Ic662eefd4058b041466d305bfabce802e76ec7df