[appmgr] Start using scheme map config file.
We replace the logic in CreateComponent that hard-codes the mapping
from component type to a branch statement with a lookup of the runner
url in the config file.

We also add an explicit "package" branch instead of making that the
default, and instead have the default path report an error about
an unrecognized scheme.

CP-69 #done

TESTED=realm_integration_test

Change-Id: I74bf3c2b747b3d2f5ac6cf7e1f097a531e74318f