[auth] Update e2e test for TokenManager.
Currently we only test with the URL of the C++ implementation, but this
change parameterizes the test to makes it easy to add testing for the
Rust implementation soon.

I've also updated the test and token manager implementation to void relying
on what I'd condider incorect behavior - calling the auth provider for
ID and Access tokens even if we don't find a refresh token to use in the
request. Happy to discuss any of these changes.

AUTH-45

Change-Id: Icd5e32749f3a5f7853260e18b9143cd27b93ef9f
Tested: CQ