[ulib][musl] Synchronize in exit
Having more than one call across the whole execution of all threads
in the entire program to either exit or quick_exit is undefined
behavior under the C standard.  So any program that allows it to
happen has a bug by definition.  But actually just synchronizing so
that all exit hooks and destructors of all sorts are reliably only
run once and the worst case is deadlock rather than mayhem is the
most useful actual behavior to have in buggy situations.

quick_exit is already fully robust with respect to (always illegal)
re-entry, always calling each hook once in the right order but
perhaps sharded across any or all threads that re-enter quick_exit.

quick_exit and exit don't interact at all, so entry of one after
the other (in the same thread via a hook or in another thread) is
as robust as re-entry of either individually.

Test: CQ
Change-Id: If0bfea49b92dc47f53e9f7dd50a47be681363eee