[build][ulib][async-loop] Define _ALL_SOURCE directly in source using extensions
The way to make the source code the most resilient to what options
it's compiled with is to define feature-test macros at the top of the
source file when a particular source file uses nonstandard extensions
declared in standard headers.

Change the build rules so that library modules exported as source get
compiled in a default feature-test environment (without _ALL_SOURCE).
This ensures the Zircon build will catch any exported source that is
not standalone with respect to its feature-test expectations.

Bug: DX-315 #done
Test: CQ
Change-Id: Ife671f2bebaa2eccbe9abbe2cd3d7ec842b9488e