[fx] use shlock instead of flock
Always use shlock, even when flock is available.
Some users on Ubuntu laptops were complaining that flock was leaving the
lock file behind, but had no trouble with shlock.
Upside: behavior on devices that have flock (Linux) and that don't have
flock (e.g. OS X) is now consistent.
Downside: shlock is not reentrant. Shouldn't be a problem for "fx build"
though, and hasn't been a problem for OS X users so far.

Change-Id: I08e4e6bdbf8dcb9af309c1ad8e858896ddc7e18c