[uapp] Add vsync mode to gfxlatency test
This adds three different vsync modes to gfxlatency, ON, OFF
and ADAPTIVE.

ON, this is standard vsync behavior. Next buffer will not be
presented until new frame is ready.

OFF, single buffer mode without vsync.

ADAPTIVE, vsync is enabled but buffer is presented even if
frame is not ready in time. No tearing is noticed as long as
frames are produced in time. If a frame takes too long to
produce, then tearing might be noticed but latency is kept
short.

TEST=gfxlatency --vsync=ON

Change-Id: I786e5d9688343f2f5cc225cf4d8a723f86048387