[ledger] Print a warning if sync-id param is missing in a benchmark.
Since we don't include the server-id in .tspec files, people trying to
use the ledger benchmarks are likely to run into this error quite often.
This adds a more descriptive error message for when --server-id flag is
missing for the benchmarks that require one. Additionally, the logic of
parsing this parameter is moved out of benchmarks into a separate helper
method.

TESTED=trace record
--spec-file=/pkgfs/packages/ledger_benchmarks/0/data/fetch.tspec

Change-Id: I47788066da230cb33bfe83e0375716889e3ff565