[async] update documentation for WaitBase::Cancel()
Note the following:
1. The class-level documentaton for WaitBase indicates that an instance
   must only be used with single-threaded dispatchers.
2. The same documentation indicates that a WaitBase instance can only
   be accessed from the dispatch thread.
3. WaitBase::Dispatch() clears the wait's pending state before yielding
   control flow to the target of the wait.

Then, if we assume that the dispatcher invokes a wait's handler
immediately after reading the wait from the port (rather than queuing
the wait for later processing), it follows that async_cancel_wait()
can never return ZX_ERR_NOT_FOUND for a wait that WaitBase::Cancel()
believes to be pending.

Update the Cancel() documentation to reflect this assumption and its
consequences, and add a corresponding runtime assertion.

Test: fx full-build (for x64)
Change-Id: Ief06c2ec8b53bddacd9289a474dabfe044f19424