[debugger] Add a skeleton symbol API.
This is part of the initial skeleton for the Symbol system.

It includes a generic "Symbol" class that's a base class for a decoded
DWARF DIE. There are some specialized subclasses for different kinds of
types. There will be more subclasses of Symbol needed in the future for
things like subroutines and local variables.

A Symbol is lazily constructed from a DIE offset stored in a LazySymbol.
This enables on-demanad decoding of the DWARF file.

These types are currently never used nor created. That will be done in
future passes.

TEST=none (not hooked up yet).

Change-Id: I4ad52245e1f140ffaf363fe4ca02065ad81db528