[tee][optee] Separate per instance device state
The OP-TEE driver needs to allow for simulatenous access from different
processes. Each instance will have it's own set of shared memory and
sessions that should be managed together. This commit splits the device
into the controller and clients. Whenever the controller device is
opened, it will create a new client device and return that instead. The
client device will handle communication with the client via the IOCTL
interface.

Test: Manually verified on Vim2 that multiple instances can be opened
and new clients will be created.

Change-Id: I018a91ae881d4339e88f5855543fc08a6c8c347b