[fidl] Keep track of the maximum handles required to represent types.
This is motivated by the desire to restrict the Clone method to types
that cannot contain handles.

This adds maximum handle tracking in TypeShape. It's mostly fairly
straightforward but there are a couple of nastier edge cases.

First unbounded vectors could have any number of members. The number of
members are in practice bounded by the maximum message size, so the
number of handles for an unbounded vector is the number of handles
multiplied by the maximum message size.

Similarly recursive structs and unions are bounded by the message size.
The compiler now has to keep track of reentrant compilation.

Messages with handles in either unbounded vectors or recursive types
will be large.

Test: wrote some simple FIDL and verified that the IR JSON was correct.

Change-Id: Ibbe2f04a63774c6cafc9205180795ccbd2d2048c