[virtcon] Try to show virtcon on all displays
This change makes the virtcon try to use all connected displays. One
display is picked as the primary display, and the image created for
that display is mirrored to all other connected displays.

Test: Boot virtcon with multiple displays; connect a new display while
the virtcon is running; unplug the primary display; unplug a secondary
display (all tests run on NUC and NUC with scaling disabled).
Change-Id: I80aeb940fa2276f6dd64649f20104c493eacfc0a