[bt][gap] Defer service discovery until after pairing
bt-host currently initiates pairing as soon as a LE connection is
created. This is temporary for non-bonded devices until support is
added to initiate pairing in response to a service request.

This patch removes a race between service discovery and pairing by
performing the former only after the pairing procedure concludes.
Services are discovered regardless of the pairing result.

NET-1144 #done
Test: 1. Use bt-le-central to pair with a keyboard.
      2. GATT services should not be displayed until after the passkey
         gets entered and the link gets encrypted.

Change-Id: Ic2666be632ccf6e59e1e1dadc91441b9ea58be2d