[display] Add display validation app
Add a test program which validates composition behavior with a usb
video capture device.

There are two significant downsides to relying on a usb video capture
device. First, it requires either USB 3.0 or decoding of a compressed
video stream (not currently implemented). Second, it puts the encoder
and the capture device in the validation pipeline, which adds unrelated
points of failure. These also potentially add unknown transformations
to the pixel data, which prevents precise validation.

This test could be significantly improved by using display capture to
memory when possible, once there are APIs to do so.

Test: self
Change-Id: Ie3d424633a93e22c91ecc54200c9e588447e06ae