[bt][rfcomm] Add MuxCommands into Frames
Frames are now aware of MuxCommands. MuxCommandFrame is added, which is
the frame type used for sending multiplexer commands along DLCI 0. This
type contains a MuxCommand; on Frame::Write(), MuxCommand::Write() is
called. Additionally, when parsing a Frame, Frame now knows how to parse
a MuxCommand into a MuxCommandFrame. Ownership of the MuxCommand can be
taken with MuxCommandFrame::TakeMuxCommand().

NET-1122 #comment
TEST=fx run-test bt-host-unittests --gtest_filter='RFCOMM_*'

Change-Id: I8a0e080820b036abf03e8dc31e3a17b5352d3305