[async] Use constant for async loop configuration.
We are going to be disallowing the default constructor for async::Loop
as well as passing NULL to async_create_loop as a config to force
the client to be explicit about whether the loop should be made the
default for the current thread. This moves all uses to use one of the
constants in zircon.

Test:CQ
Change-Id: I0a8e569df9b6519289bc11782a4ab6d099e833d6