[async] Use constant for async loop configuration.
This change enforces that async_create_loop() can not take a NULL config
and removes the default null_ptr config from async::Loop()

A couple of client uses in zircon had to be updated as well.

The change needs to wait for the depedent changes other layers have
cleared the CQ. Higher level trybots will be run first.

Test:CQ
Change-Id: Ia643da71acad0d843297b4dbbad5cdf6db2bb741