[module_context] Add ModuleContext.Done
ModuleContext.Done allows modules to inform the framework that they
want to be stopped.

In the future it is possible that ModuleContext.Done will delegate
responsibility for stopping to the holder of its module controller, if
it has registered a handler for it.

A watcher interface would be added to ModuleController, and the holder
of a ModuleController could register a watcher for the done signal. This
would allow the holder of the ModuleController to do things like animate
the stopping explicitly. If no such watchers were registered the
framework would still stop the module.

Even if we decide to go with the more involved solution above in the
future, the behavior would only change when a watcher is registered,
meaning that existing calls to ModuleContext.Done would continue to
behave as expected.

In addition, this also converts StoryProvider.RunningStories to use
the operation queue.

TESTED=tests/module_context

MI4-985 #done

Change-Id: I32a736fee46b48b20bc877554aa280b203b870ac