[blobfs] add support for depfile output
depfile is a GN & ninja facility for tracking build target inputs that are
not otherwise declared in GN action "inputs" or "sources". During the GN
build of the blobfs images, the actions produce a manifest file that lists
all of the contents of the blob image, but do not list the merkleroots of
those files. As such, ninja does not determine that there is a need to
rebuild the blobfs unless it has some other file that declares the contents
of the manifest are inputs to the blobfs output file. Producing a depfile is
a normal solution to this problem.

Test: manual. Added depfile parameter to the build rule, ran the build and
observed correct depfile output. Modified a blobfs input and observed ninja
rebuild the blobfs target.

ZX-2378

Change-Id: Idd7500c6bbc8d798e185caf3147a527f8753e827