[async] update documentation for TaskBase::Cancel()
Note the following:
1. The class-level documentaton for TaskBase indicates that an instance
   must only be used with single-threaded dispatchers.
2. The same documentation indicates that a TaskBase instance can only
   be accessed from the dispatch thread.
3. TaskBase::Dispatch() clears the wait's pending state before yielding
   control flow to the target of the wait.

Then, if we assume that the dispatcher invokes a task's handler
immediately after removing the task from the task list, it follows
that async_cancel_task() can never return ZX_ERR_NOT_FOUND for a wait
that TaskBase::Cancel() believes to be pending.

Note that, async_loop_t does not _literally_ fit the above assumption,
since async_loop_dispatch_tasks() moves the ready tasks from the
|task_list| to the |due_list| before invoking any handlers. However,
async_loop_t _effectively_ fits the above assumption, because:
a) async_loop_cancel_task() behaves the same regardless of whether the
   task is in the |task_list| or the |due_list|, and
b) even if a) were not true, with a single-threaded instance of
   async_loop_t, and TaskBase only being accessed from the dispatch
   thread (see Note #2 above), async_loop_cancel_task() can not be
   called between the time when the task is removed from |task_list|,
   and the time when the handler is invoked.

Given this, update the Cancel() documentation to reflect the
simplifying assumption and its consequences, and add a corresponding
runtime assertion.

Test: fx full-build (on x64)
Change-Id: I18ca7fe8befedfccfa2dfe76433e30cd9b6d2768