[debugger] Add more symbol infrastructure.
Adds a Symbol subclasses for a Function and hooks up parsing code for
that, BaseType, and ModifiedType.

A reference to the Function was added to the Location object which is in
use for other things in the project. This is not currently hooked up,
however.

The function base a base class of CodeBlock which I plan to use for
lexical_scope and inlined subroutines in the future.

Functionality was added to the DWARF DIE decoder to parse more stuff,
notably DIE cross-references and file name references, both of which are
automatically resolved.

TEST=unit

Change-Id: Ifa05386504b5c909c6308d7d7167364d44214b28