[debugger] Support for variable sized registers.
Wrapped over the register information into a class that encapsulates the
value handling of the underlying information. This will be extended to
also contains the sub-register information in a future CL.

This wrappers hook up to the register and track whether their
information is up to date.

Removed client facing naming in the debug_ipc and replaced with static
information. Moved the naming to the client instead on relying on the
debug agent to supply that information.

TEST=unit, manual

Change-Id: Id79797d74c02e4a73137cb1498e6316ac60cadf9