[media] fix MediaType readability
This CL replaces MediaType with StreamType. Prefixing types in
fuchsia.media with 'Media' is prohibited, and fuchsia.media.Type
isn't specific enough, hence StreamType.

This change also makes the medium-specific types more useable
independent of StreamType itself. These types are useful when the
medium and encoding of a stream is known. This allows us to replace
AudioPcmFormat with AudioStreamType (which is also the replacement
for AudioMediaTypeDetails) eliminating some duplication.

StreamType dispenses with the Medium enum, replacing it with the tag
on the union MediumSpecificStreamType.

AudioStreamType now has a channel configuration field, which we've
needed for awhile.

TEST: No behavior change. CQ regression testing.
MTWN-93 #done

Change-Id: Ie9f4e9cab08c371319f5b51bbb3e80c93e98dec2