[amlogic-video] Fix parsing small chunks of video.
The previous approach of using the fetch complete interrupt wasn't
working if the input was given in small chunks - either garbage was
being read or some input was being thrown away; it's unclear which.

Instead, parse a buffer with the ES search pattern (0, 0, 1) after the
main buffer. We can then use the interrupt for when the start code is
found as a signal that the buffer is parsed. The parser fifo also seems
to need to be cleared after reading to avoid getting spurious
interrupts.

Also add a test that splits the h.264 file into NAL units and parses
them individually.

TEST: amlogic_decoder_tests=true

Change-Id: I42670eaaa6a702093e3dc7c69390a561b68f0d8f