[rust] Fix race condition in on_signals.
This will fix bug due to which logger was not printing kernel logs when
large amount of logs were produced.

The race condition:
- we have two threads t1 and t2
- t2 is wating on port
- t1 creates OnSignals object which adds a wait_async to common port
- t2 wakes up to just added wait and calls set_signals of OnSignalsReceiver
- t1 tries to poll on OnSignals and calls get_signals of OnSignalsReceiver
- t1 reads maybe_signal as 0
- t2 sets maybe_signal to signal recieved from zircon and calls
task.wake
- task was empty so signal got lost
- t1 registers to task but now there are no more signals to wake task up
- client never recieves any signal

TEST: tested using
https://fuchsia-review.googlesource.com/c/garnet/+/175850 which produces
endless kernel logs. Before the fix the program was hanging as soon as
it was executed. after the fix it did not hang for 30 min.

CP-101 #done
US-517 #done

Change-Id: Ib71686f16187495c65a8a6110820f295684a0b8f