[modular][future] What's taking so long?
Adding WaitWithTimeout, which allows for a callback, diagnostic, and
cancellation after a timeout when waiting on a bunch of futures.

This also stops Wait from holding onto its component futures. Instead,
the future returned by Wait will be released if all its component
futures have either been completed or destroyed. The justification is
that if any component future is destroyed, it cannot possibly be
completed and so the Wait would never be completed either.

Test: fx run-test peridot_public_lib_unittests
      fx shell /pkgfs/packages/modular_tests/0/test/run_modular_tests.sh
      usage in downstream code
Change-Id: I7479cf04c7f99b99727d4f3c53d9cd807cc40c8f