[audio] DFX library and test
This CL adds a sample implementation of a DFX shared library, and an
example set of tests to validate that library. Both sets of code are
targeted at system builders who expect to create device effects.

The library implements three effects, in the binary libaudio_dfx.so.
The tests of that shared library are contained in audio_dfx_tests.

Test: this CL introduces new test binary 'audio_dfx_tests'

Change-Id: I6535504fc07f7ff30341e6ebd4aefd74553ea9fb