[arm64][hypervisor] Save and restore GIC APR register
The APR register controls the active priority of interrupt controller,
and should not be inconsistent with the state of the GIC. Prior to this
patch it was possible for the value of the APR from a previously
terminated guest (or theoretically another guest running in parallel on
the same CPU) to leak over to another guest. If the previous guest
terminated while APR was set to 1, then priority 0 interrupts would be
considered spurious and the guest would get stuck in a loop.

TEST=Ran guest_integration_tests on vim2. Ran zircon_guest on qemu to
test GICv3.

Change-Id: Iabf439eadb1a3c4a703b51590c29a2a9d6e35755