[cloud_provider_firestore] track existence via placeholder documents
This patch adds logic to make a best-effort attempt to create an
explicit placeholder document at each path where we currently have an
implicit empty document.

This is to workaround limitations of the web client API, which excludes
implicit documents from query results, for the purposes of supporting
deletion in the development cloud dashboard.

LE-522 #fixed

Test: cloud_provider_firestore_unittests
Change-Id: Idc180dfcee72681f2c33443a0928225766521790