[scenic] Add gn arg to ignore vsync.
Due to the way frames are scheduled, the framerate will evenly divide
the screen refresh rate (60fps). With "scenic_ignore_vsync=true" as a gn
arg, scenic will try to draw new frames as soon as possible, allowing
the framerate to precisely reflect how long it took to render.

This is useful for benchmarking, as then small changes in rendering
performance are more likely to be reflected in the rendered framerate.

Test: "set_root_view vkcube_image_pipe_swapchain"

Change-Id: Ia4806db8ee76a87b33cd6ea11e6cb4c97b10674f