[serve-updates] Don't use SSH tunnel when --direct
The first iteration of "serve-updates --direct" no longer used the SSH
tunnel as a transport, but it still used it to help detect device
reboots. This change removes the SSH tunnel entirely when this flag is
used and instead uses ping to detect device reboots. Also, when the
direct flag is used, serve-updates will only attempt to re-discover the
target's address when a ping fails.

To facilitate debugging, the verbose flag can now be specified twice to
enable additional logging (such as every netaddr call and ping attempt).

Test=manual

Change-Id: I7c7a94fd1c5518126af96083ae40e52b6727f7a1