[benchmarks] Disable broken leder.add_new_page test
Right now, perf builders only show a breakage if something
goes horribly wrong and the entire build fails.  If an individual
test fails (which ledger.add_new_page is), the build stays green
and the only indication that the test is failing is the red step
on the build log page. No email or anything.

Soon I'll change the perf builders so that we report a breakage for the
entire build when an individual test fails, so we should disable broken
tests and attach bugs to fix them.

Additionally, we'll need some way to report individual test
breakages moving forward.

Test: On CQ

Change-Id: I6ffdf257ee862902fc0b6ab8f9ed9b7ae7dd1279