[appmgr][components] Launch binary as specd by cmx
root_loader now puts fuchsia-pkg:// urls in resolved_url when a .cmx is
requested to be launched. Compare that to a file:// url for a launch
from package name, which points to the package dir.

When we read a .cmx in realm, we check whether package.resolved_url is
fuchsia-pkg://. If so, then we read the correct cmx as requested. If
not, we default to the <package_name>.cmx.

Finally, wire up the information of the ELF binary location specified in
the cmx and launch that appropriately. The result is that we have
multiple ELF binary components per package.

CP-19 #comment

TESTED=
run -d fuchsia-pkg://fuchsia.com/component_hello_world#meta/hello_world.cmx
run -d fuchsia-pkg://fuchsia.com/component_hello_world#meta/hello_world2.cmx
run -d component_hello_world
/pkgfs/packages/appmgr_unittests/0/test/appmgr_unittests

Change-Id: If7bfa76de90da89a38061187383c2b74614745eb