[cp] Add a generic JSON parser class.
JSONParser is a lightweight class that takes as input a path to a
JSON file and provides the client a rapidjson::Document. The class
also integrates error reporting, which supports line numbers on parse
errors.

Clients can use this class to implement their own JSON parsers. Next
step is to refactor the various JSON parsers in core platform to
subclass this class.

TESTED=json_document_unittest

Change-Id: Ia01a63a72e247226a5725bc537581b0e2c33a3ed