[products] introduce "products" to describe builds
A "product" definition is stored under {layer}/packages/products. A product
definition is a JSON file containing and object with the following fields:

monolith:
  a list of packages included in OTA images, base system images, and the
  distribution repository.
preinstall:
  a list of packages pre-installed on the system (also added to the
  distribution repository)
available:
  a list of packages only added to the distribution repository)

Note: preinstall does not fully work as yet. The blobs are added to the
blobfs, however, no dynamic index is added to the minfs image that would
indicate to pkgsvr that they are preinstalled. This feature requires changes
in Garnet. Having the blobs pre-seeded in blobfs is still useful as it may
greatly speed up time to activation of those packages.

Test: tested against I4e797db6235bf7a2ec78a34b636ba01da735e417

Change-Id: I8eb6d84343ea0b047dbf101b5b26e78bac361648