[build][zbi] Exclude modules from BOOTFS by group
This adds a negation feature to the --groups switch to zbi: if the
group list is preceded by a `!` then it excludes the listed groups
and allows all others (and unmarked entries).  Incidentally it also
fixes group matching (really, parsing the option value), which was
completely broken since it wasn't being tested in the build before.

The build now uses this to control the groups that go into the
zircon.zbi BOOTFS image via the USER_MANIFEST_GROUPS makefile
variable.  The default setting is `!sample,ddk-sample` to include
everything except modules marked `sample` or `ddk-sample`.

Bug: ZX-2380 #done
Test: CQ and manual inspection with zbi -tv
Change-Id: I41d651774f607acaf97c5fc47c88b2ea8dddd923