[fuchsia.guest] Handle creation by vsock origin
From the host, allow the process that is connecting or accepting to
provide the handle with which to transmit and receive data.

From the guest, defer all handle creation to host processes.

This moves the handle creation outside of the guestmgr, with the
exception of guest to guest communication.

This allows us to use sockets or channels where appropriate, so that we
can wire up guests to FIDL services.

Test: Added new tests, and ran "guest socat".
Change-Id: I8d427f329573248506bcab6acbd91f5ec0dd5360