[debugger] Return function objects for locations.
Replaces the function name in the Location object with the object Symbol
representing the full function symbol.

Add support for namespace symbols.

Move the enclosing symbol context to every symbol type so the dependency
tree can be walked up consistently across symbols.

Move the "assigned name" to the root Symbol object since so many Symbols
end up having a name and this simplifies the special-casing required.

Add helper functions to qualify the scopes based on structure and
namespace hierarchy. This works for types and other members. This is
separate from the Symbol class because I expect we'll want some
different options. As an example, I want to try to highlight the
function name separately from the namespaces and template goop in the
backtrace listing, and to eliminate namespace scoping for objects when
you're already inside that namespace. This desire informed the current
implementation that separates out computation of the hierarchy from
generating the qualifier string, from generating the fully-qualified
name.

Fix a small error in the debug_agent where a return value wasn't checked.

TEST=unit

Change-Id: I0c424e4ba06a1d20d9a1e391623bfdb97dd6e54e