[fostr] add ability to generate ostream formatter code
This CL adds the ability to generate ostream formatter code based on
FIDL definitions. Two examples are given: one for fuchsia.mem, which
is defined in zircon, and one for fuchsia.sys, which is defined in
garnet. This approach has been verified with fuchsia.mediaplayer and
all the things it depends on.

Adding a new FIDL library is simple: just add a new fuchsia.xxx
subdirectory with a BUILD.gn like the fuchsia.sys example. The target
must depend on the fostr/fidl targets for the libraries on which
fuchsia.xxx depends.

If a library has hand-rolled formatters, they can be accommodated as
well with a couple of BUILD.gn files and a formatting.h that includes
the handle-rolled formatting.h.

Includes look like this:

include "lib/fostr/fidl/fuchsia/sys/formatting.h"

The targets look like this:

//garnet/public/lib/fostr/fidl/fuchsia.sys

TEST: Tested manually with fuchsia.mediaplayer
Change-Id: Ife62fa838deb364fea360077915445b1ac84b994