[trace] Output a "test_suite" field in the JSON test results file
Currently the test suite name (as used by the Catapult dashboard) is
taken from the name of the JSON file (e.g. "benchmark_example.json" is
given the test suite name "fuchsia.benchmark_example").  However, the
new Catapult converter in Garnet takes the test suite name from the
test_suite field in the JSON file.  Change the "trace" tool to be
capable of outputting that field.

Bug: IN-444
Test: manual: run 'trace' on benchmark_example and check the output

Change-Id: I659bd840cdefdc72b2638464d065c350644987bf